+++ tests/test_stream.py
@@ -9,6 +9,7 @@ import os
 import sys
 import unittest
 from io import StringIO
+
 from pyspartn.exceptions import SPARTNMessageError, SPARTNParseError, ParameterError
 from pyspartn.spartnreader import SPARTNReader, SPARTNMessage
 from pyspartn.spartntypes_core import ERRRAISE, ERRIGNORE, ERRLOG
@@ -198,7 +199,10 @@ class StreamTest(unittest.TestCase):
         with self.assertRaisesRegex(ParameterError, EXPECTED_ERROR):
             i = 0
             spr = SPARTNReader(
-                self.streamSPARTN, quitonerror=ERRRAISE, decrypt=True, key=None
+                self.streamSPARTN,
+                quitonerror=ERRRAISE,
+                decrypt=True,
+                key=None,
             )
             for raw, parsed in spr:
                 if raw is not None: